unit10/requirements.txt
python-dotenv==1.1.0
llama-index==0.12.35
llama-index-core==0.12.35
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.38
openai==1.78.0
gradio==5.29.0
gradio_client==1.10.0 